hue lights count response processing works

+++ wifi/hue_lights_client.h
@@ -26,7 +26,7 @@ private:
   bool displayConnectionDetails();
   bool httpRequest(const __FlashStringHelper* method, char* url, char* headers, char* body);
   uint16_t readHTTPResponseStatus();
-  uint16_t readHTTPLightsResponse();
+  uint8_t readHTTPLightsResponse();
 
 private:
